исправление багов и рефакторинг

+++ spu/src/spu.h
@@ -112,11 +112,11 @@ void print_spu_error(SPUStatus err);
 
 void print_spu_verify_res_(int verify_res);
 
-SPUStatus run_program(SPU *spu_ptr, int *prog_res);
+SPUStatus run_program(SPU *spu_ptr, double *prog_res);
 
 //! @brief Executes command at which ip points to
 //! and moves ip.
-SPUStatus exec_curr_cmd_(SPU *spu_ptr, int *prog_res);
+SPUStatus exec_curr_cmd_(SPU *spu_ptr, double *prog_res);
 
 //! @note BE VERY CAREFUL USING THIS! MAYBE NOT ALL STACK
 //! FUNCTIONS RETURN STACK_ERROR, ALSO BE AWARE THAT
